chore(ConsoleVerbosity): Move `ConsoleVerbosity` to  `toolr.utils._console`
