+++ tests/context/test_output.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from unittest import mock
 
-from toolr._context import ConsoleVerbosity
 from toolr._context import Context
 from toolr.utils._console import Consoles
+from toolr.utils._console import ConsoleVerbosity
 
 
 def test_debug_output(parser, repo_root):
