+++ tests/utils/test_console.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import pytest
 
-from toolr._context import ConsoleVerbosity
 from toolr.utils._console import Consoles
+from toolr.utils._console import ConsoleVerbosity
 
 
 @pytest.fixture(
